requirements.txt
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
fastapi==0.124.2
h11==0.16.0
idna==3.11
psycopg2-binary==2.9.11
pydantic[email]==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
alembic==1.17.2
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
